requirements.txt
langchain==1.0.0
langchain-google-genai
langchain-community
langchain-text-splitters
python-dotenv
streamlit
chromadb
wikipedia
